+++ src/core/command/command.js
@@ -110,6 +110,10 @@ export default class Command {
                     combined[combined.length - 1] = combined[combined.length - 1].slice(1)
                     inString = true
                 }
+                if (/["'“”‘’]$/.test(i)) {
+                    combined[combined.length - 1] = combined[combined.length - 1].slice(0, -1)
+                    inString = false
+                }
             }
         }
         command = combined