+++ src/modules/utility/commands/index.js
@@ -1,2 +1,3 @@
 export roll from './roll'
 export test from './test'
+export batch from './batch'